requirements.txt
flask==2.2.5
flask-cors==3.0.10
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==0.21.1
newsapi-python==0.2.7
SQLAlchemy==2.0.28
pytest==7.4.4
gunicorn==21.2.0
selenium==4.21.0
rapidfuzz==3.6.2
serpapi==0.1.5
torch==1.13.1
torchvision==0.14.1
opencv-python==4.9.0.80
numpy==1.24.4
Pillow==10.3.0
tensorflow==2.2.0
imageio==2.9.0
ffmpeg-python==0.2.0
face_recognition==1.3.0 